pyproject.toml
[tool.poetry]
name = "fs-test-debug"
version = "0.1.0"
description = ""
authors = ["Armen Mkrtchyan <amkrtchyan@techdaisy.co>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
flask = "^2.3.3"
flask-sqlalchemy = "^3.1.1"
flask-security-too = "^5.3.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
